Move xkeysyms module into x11 backend
The data in this xkeysyms.py is x11-specific and so belongs to the x11
backend. The only place outside of the backend where it is used is in
the prompt widget, however this widget will soon be updated to work with
other backends, requiring it to use another means of binding its methods
to key codes.
